Falcon-40B: speed up native reference load (no_init_weights + bf16 for demo)
The demo loads the full 60-layer reference, and the native loader was randomly
initialising all ~40B params before overwriting them and converting the whole
model to float32 — together slow enough to exceed the demo's e2e step timeout
(weight load alone took ~7.5 min; the fix loads correctly but timed out).

- Construct under transformers' no_init_weights() so the about-to-be-overwritten
  random init is skipped (degrades to a no-op context if the symbol moves).
- Add a dtype arg (default float32 for the PCC tests, which feed float32 inputs
  and need it to avoid mixed-dtype LayerNorm); the demo passes bfloat16 — it only
  feeds token ids, and the e2e demo runs perf_mode (no output assertion), so
  bf16 halves host memory and skips the slow float32 conversion.

+++ models/demos/t3000/falcon40b/demo/demo.py
@@ -237,8 +237,9 @@ def run_falcon_demo_kv(
 
     # transformers 5.x silently fails to populate the vendored Falcon via from_pretrained (the TT
     # model then gets unloaded weights and the demo hangs on device); load the raw checkpoint in
-    # its native layout instead. See issue #47924.
-    hugging_face_reference_model = load_falcon_reference_model(model_version)
+    # its native layout instead. bfloat16 matches the checkpoint dtype and keeps the full-model
+    # load within the demo's step timeout. See issue #47924.
+    hugging_face_reference_model = load_falcon_reference_model(model_version, dtype=torch.bfloat16)
     state_dict = hugging_face_reference_model.state_dict()
 
     profiler.end("loading_weights")
